Split according to surveys
