+++ fink_filters/ztf/filter_dwarf_agn/filter.py
