+++ fink_filters/ztf/filter_dwarf_agn/list_dwarfs_AGN_RADEC.txt
