+++ fink_filters/ztf/filter_early_sn_candidates/filter.py
