+++ fink_filters/ztf/filter_known_tde/data/TDElist_Gezari2021.tsv
