+++ fink_filters/ztf/filter_known_tde/data/Table1_Hammerstein
