+++ fink_filters/ztf/filter_magnetic_cvs/filter.py
