+++ fink_filters/ztf/filter_quality_cuts/filter.py
