+++ fink_filters/ztf/filter_simbad_candidates/filter.py
