+++ fink_filters/ztf/filter_sn_candidates/filter.py
