+++ fink_filters/ztf/filter_symbiotic_stars/data/cataclysmic_variables_catalog.txt
