+++ fink_filters/ztf/filter_symbiotic_stars/data/symbiotic_stars_catalog.txt
