+++ fink_filters/ztf/filter_yso_spicy_candidates/filter.py
